Properly fix lack of disco plugin
That was quite a ride!
Fixes Error: Missing strophe-plugins (disco plugin is required)!
With Vite.
The plugin is compiled in a way that doesn't work with Vite, so the code to register the plugin doesn't run. That's my guess anyway.
Why this 200-line plugin isn't included with lib-jitsi-meet like (all?) other plugins are is also beyond me.
But hey, we have it working with Vite now! Yay!